Sun, Aug 11, 2024, 11:42 AM +02:00

+++ readability/readability.c
@@ -19,8 +19,8 @@ int main(void)
     int sentences = count_sentences(text);
 
     // Compute the Coleman-Liau index
-    float l = (letters / words) * 100;
-    float s = (sentences / words) * 100;
+    float l = ((float) letters / words) * 100;
+    float s = ((float) sentences / words) * 100;
     int index = 0.0588 * l - 0.296 * s - 15.8;
     // Print the grade level
     printf("%d\n", letters);
